requirements.txt
pandas==2.1.3
numpy==1.24.3
scikit-learn==1.3.2
matplotlib==3.8.2
seaborn==0.13.0
tqdm==4.66.1
fastapi==0.104.1
uvicorn[standard]==0.24.0
streamlit==1.28.1
plotly==5.18.0
Pillow==10.1.0
requests==2.31.0
python-multipart==0.0.6